refactor: core 로직은 lib/, UI 컴포넌트는 components/로 분리
